Fix initializer lists compilation issues with Clang
src/Dialogs/ModifyOffsetsDialog.cpp:121:11: error: non-constant-expression cannot be narrowed from type 'long' to 'int' in initializer list [-Wc++11-narrowing]
src/Dialogs/ModifyOffsetsDialog.cpp:121:14: error: non-constant-expression cannot be narrowed from type 'long' to 'int' in initializer list [-Wc++11-narrowing]
src/MapEditor/Renderer/Overlays/LineInfoOverlay.cpp:61:48: error: type 'double' cannot be narrowed to 'int' in initializer list [-Wc++11-narrowing]
src/MapEditor/Renderer/Overlays/ThingInfoOverlay.cpp:63:86: error: type 'double' cannot be narrowed to 'int' in initializer list [-Wc++11-narrowing]

+++ src/MapEditor/Renderer/Overlays/ThingInfoOverlay.cpp
@@ -60,7 +60,7 @@ EXTERN_CVAR(Bool, use_zeth_icons)
 // -----------------------------------------------------------------------------
 // ThingInfoOverlay class constructor
 // -----------------------------------------------------------------------------
-ThingInfoOverlay::ThingInfoOverlay() : text_box_{ "", Drawing::Font::Condensed, 100, 16 * (Drawing::fontSize() / 12.0) }
+ThingInfoOverlay::ThingInfoOverlay() : text_box_{ "", Drawing::Font::Condensed, 100, int(16 * (Drawing::fontSize() / 12.0)) }
 {
 }
